More cleanup and move files.
