+++ src/TcecEvaluationBot.ConsoleUI/Services/ChessPosDbProxy.cs
@@ -707,353 +707,4 @@
             stream.Write(bytes, 0, bytes.Length);
         }
     }
-
-    public enum MergeMode
-    {
-        None,
-        Consecutive,
-        Any,
-    }
-
-    public class DatabaseMergableFile
-    {
-        public DatabaseMergableFile(JObject json)
-        {
-            this.Name = json["name"].Value<string>();
-            this.Size = json["size"].Value<ulong>();
-        }
-
-        public string Name { get; private set; }
-
-        public ulong Size { get; private set; }
-    }
-
-    public class DatabaseSingleLevelStats
-    {
-        public DatabaseSingleLevelStats(JObject json)
-        {
-            this.NumGames = json["num_games"].Value<ulong>();
-            this.NumPositions = json["num_positions"].Value<ulong>();
-            this.TotalWhiteElo = json["total_white_elo"].Value<ulong>();
-            this.TotalBlackElo = json["total_black_elo"].Value<ulong>();
-            this.NumGamesWithElo = json["num_games_with_elo"].Value<ulong>();
-            this.NumGamesWithDate = json["num_games_with_date"].Value<ulong>();
-
-            if (this.NumGamesWithElo > 0)
-            {
-                this.MinElo = json["min_elo"].Value<ulong>();
-                this.MaxElo = json["max_elo"].Value<ulong>();
-            }
-
-            if (this.NumGamesWithDate > 0)
-            {
-                this.MinDate = Date.FromString(json["min_date"].Value<string>(), '-');
-                this.MaxDate = Date.FromString(json["max_date"].Value<string>(), '-');
-            }
-        }
-
-        public DatabaseSingleLevelStats(DatabaseSingleLevelStats other)
-        {
-            this.NumGames = other.NumGames;
-            this.NumPositions = other.NumPositions;
-            this.TotalWhiteElo = other.TotalWhiteElo;
-            this.TotalBlackElo = other.TotalBlackElo;
-            this.NumGamesWithElo = other.NumGamesWithElo;
-            this.NumGamesWithDate = other.NumGamesWithDate;
-            this.MinElo = other.MinElo;
-            this.MaxElo = other.MaxElo;
-            this.MinDate = other.MinDate;
-            this.MaxDate = other.MaxDate;
-        }
-
-        public ulong NumGames { get; private set; }
-
-        public ulong NumPositions { get; private set; }
-
-        public ulong TotalWhiteElo { get; private set; }
-
-        public ulong TotalBlackElo { get; private set; }
-
-        public ulong NumGamesWithElo { get; private set; }
-
-        public ulong NumGamesWithDate { get; private set; }
-
-        public ulong MinElo { get; private set; }
-
-        public ulong MaxElo { get; private set; }
-
-        public Date MinDate { get; private set; }
-
-        public Date MaxDate { get; private set; }
-
-        public void Add(DatabaseSingleLevelStats other)
-        {
-            this.NumGames += other.NumGames;
-            this.NumPositions += other.NumPositions;
-            this.TotalWhiteElo += other.TotalWhiteElo;
-            this.TotalBlackElo += other.TotalBlackElo;
-
-            if (this.NumGamesWithElo == 0)
-            {
-                this.MinElo = other.MinElo;
-                this.MaxElo = other.MaxElo;
-            }
-            else if (other.NumGamesWithElo != 0)
-            {
-                this.MinElo = Math.Min(this.MinElo, other.MinElo);
-                this.MaxElo = Math.Min(this.MaxElo, other.MaxElo);
-            }
-
-            if (this.NumGamesWithDate == 0)
-            {
-                this.MinDate = other.MinDate;
-                this.MaxDate = other.MaxDate;
-            }
-            else if (other.NumGamesWithDate != 0)
-            {
-                this.MinDate = Date.Min(this.MinDate, other.MinDate);
-                this.MaxDate = Date.Max(this.MaxDate, other.MaxDate);
-            }
-
-            this.NumGamesWithElo += other.NumGamesWithElo;
-            this.NumGamesWithDate += other.NumGamesWithDate;
-        }
-    }
-
-    public class DatabaseSingleLevelImportStats : DatabaseSingleLevelStats
-    {
-        public DatabaseSingleLevelImportStats(JObject json)
-            : base(json)
-        {
-            this.NumSkippedGames = json["num_skipped_games"].Value<ulong>();
-        }
-
-        public DatabaseSingleLevelImportStats(DatabaseSingleLevelImportStats other)
-            : base(other)
-        {
-            this.NumSkippedGames = other.NumSkippedGames;
-        }
-
-        public ulong NumSkippedGames { get; private set; }
-
-        public void Add(DatabaseSingleLevelImportStats other)
-        {
-            base.Add(other);
-            this.NumSkippedGames += other.NumSkippedGames;
-        }
-    }
-
-    public class DatabaseStats
-    {
-        public DatabaseStats(JObject json)
-        {
-            this.StatsByLevel = new Dictionary<GameLevel, DatabaseSingleLevelStats>
-            {
-                { GameLevel.Engine, new DatabaseSingleLevelStats(json["engine"].Value<JObject>()) },
-                { GameLevel.Human, new DatabaseSingleLevelStats(json["human"].Value<JObject>()) },
-                { GameLevel.Server, new DatabaseSingleLevelStats(json["server"].Value<JObject>()) },
-            };
-        }
-
-        public Dictionary<GameLevel, DatabaseSingleLevelStats> StatsByLevel { get; private set; }
-
-        public DatabaseSingleLevelStats GetTotal()
-        {
-            DatabaseSingleLevelStats total = new DatabaseSingleLevelStats(this.StatsByLevel[GameLevel.Engine]);
-            total.Add(this.StatsByLevel[GameLevel.Human]);
-            total.Add(this.StatsByLevel[GameLevel.Server]);
-            return total;
-        }
-    }
-
-    public class DatabaseImportStats
-    {
-        public DatabaseImportStats(JObject json)
-        {
-            this.StatsByLevel = new Dictionary<GameLevel, DatabaseSingleLevelImportStats>
-            {
-                { GameLevel.Engine, new DatabaseSingleLevelImportStats(json["engine"].Value<JObject>()) },
-                { GameLevel.Human, new DatabaseSingleLevelImportStats(json["human"].Value<JObject>()) },
-                { GameLevel.Server, new DatabaseSingleLevelImportStats(json["server"].Value<JObject>()) },
-            };
-        }
-
-        public Dictionary<GameLevel, DatabaseSingleLevelImportStats> StatsByLevel { get; private set; }
-
-        public DatabaseSingleLevelImportStats GetTotal()
-        {
-            DatabaseSingleLevelImportStats total = new DatabaseSingleLevelImportStats(this.StatsByLevel[GameLevel.Engine]);
-            total.Add(this.StatsByLevel[GameLevel.Human]);
-            total.Add(this.StatsByLevel[GameLevel.Server]);
-            return total;
-        }
-    }
-
-    public class DatabaseManifest
-    {
-        public DatabaseManifest(JObject json)
-        {
-            this.Schema = json["schema"].Value<string>();
-            this.Version = json["version"].Value<string>();
-        }
-
-        public string Schema { get; private set; }
-
-        public string Version { get; private set; }
-    }
-
-    public class DatabaseSupportManifest
-    {
-        public DatabaseSupportManifest(JObject json)
-        {
-            this.SupportedExtensions = new List<string>();
-
-            foreach (var ext in json["supported_file_types"].Value<JArray>())
-            {
-                this.SupportedExtensions.Add(ext.Value<string>());
-            }
-
-            switch (json["merge_mode"].Value<string>())
-            {
-                case "none":
-                    this.MergeMode = MergeMode.None;
-                    break;
-
-                case "consecutive":
-                    this.MergeMode = MergeMode.Consecutive;
-                    break;
-
-                case "any":
-                    this.MergeMode = MergeMode.Any;
-                    break;
-            }
-
-            this.MaxGames = json["max_games"].Value<ulong>();
-            this.MaxPositions = json["max_positions"].Value<ulong>();
-            this.MaxInstancesOfSinglePosition = json["max_instances_of_single_position"].Value<ulong>();
-
-            this.HasOneWayKey = json["has_one_way_key"].Value<bool>();
-            if (this.HasOneWayKey)
-            {
-                this.EstimatedMaxCollisions = json["estimated_max_collisions"].Value<ulong>();
-                this.EstimatedMaxPositionsWithNoCollisions = json["estimated_max_positions_with_no_collisions"].Value<ulong>();
-            }
-
-            this.HasCount = json["has_count"].Value<bool>();
-
-            this.HasEloDiff = json["has_elo_diff"].Value<bool>();
-            if (this.HasEloDiff)
-            {
-                this.MaxAbsEloDiff = json["max_abs_elo_diff"].Value<ulong>();
-                this.MaxAverageAbsEloDiff = json["max_average_abs_elo_diff"].Value<ulong>();
-            }
-
-            this.HasWhiteElo = json["has_white_elo"].Value<bool>();
-            this.HasBlackElo = json["has_black_elo"].Value<bool>();
-            if (this.HasWhiteElo || this.HasBlackElo)
-            {
-                this.MinElo = json["min_elo"].Value<ulong>();
-                this.MaxElo = json["max_elo"].Value<ulong>();
-                this.HasCountWithElo = json["has_count_with_elo"].Value<bool>();
-            }
-
-            this.HasFirstGame = json["has_first_game"].Value<bool>();
-            this.HasLastGame = json["has_last_game"].Value<bool>();
-
-            this.AllowsFilteringTranspositions = json["allows_filtering_transpositions"].Value<bool>();
-            this.HasReverseMove = json["has_reverse_move"].Value<bool>();
-
-            this.AllowsFilteringByEloRange = json["allows_filtering_by_elo_range"].Value<bool>();
-            this.EloFilterGranularity = json["elo_filter_granularity"].Value<ulong>();
-
-            this.AllowsFilteringByMonthRange = json["allows_filtering_by_month_range"].Value<bool>();
-            this.MonthFilterGranularity = json["month_filter_granularity"].Value<ulong>();
-
-            this.MaxBytesPerPosition = json["max_bytes_per_position"].Value<ulong>();
-
-            if (json.ContainsKey("estimated_average_bytes_per_position"))
-            {
-                this.EstimatedAverageBytesPerPosition = Optional<ulong>.Create(json["estimated_average_bytes_per_position"].Value<ulong>());
-            }
-            else
-            {
-                this.EstimatedAverageBytesPerPosition = Optional<ulong>.CreateEmpty();
-            }
-        }
-
-        public IList<string> SupportedExtensions { get; private set; }
-
-        public MergeMode MergeMode { get; private set; }
-
-        public ulong MaxGames { get; private set; }
-
-        public ulong MaxPositions { get; private set; }
-
-        public ulong MaxInstancesOfSinglePosition { get; private set; }
-
-        public bool HasOneWayKey { get; private set; }
-
-        public ulong EstimatedMaxCollisions { get; private set; }
-
-        public ulong EstimatedMaxPositionsWithNoCollisions { get; private set; }
-
-        public bool HasCount { get; private set; }
-
-        public bool HasEloDiff { get; private set; }
-
-        public ulong MaxAbsEloDiff { get; private set; }
-
-        public ulong MaxAverageAbsEloDiff { get; private set; }
-
-        public bool HasWhiteElo { get; private set; }
-
-        public bool HasBlackElo { get; private set; }
-
-        public ulong MinElo { get; private set; }
-
-        public ulong MaxElo { get; private set; }
-
-        public bool HasCountWithElo { get; private set; }
-
-        public bool HasFirstGame { get; private set; }
-
-        public bool HasLastGame { get; private set; }
-
-        public bool AllowsFilteringTranspositions { get; private set; }
-
-        public bool HasReverseMove { get; private set; }
-
-        public bool AllowsFilteringByEloRange { get; private set; }
-
-        public ulong EloFilterGranularity { get; private set; }
-
-        public bool AllowsFilteringByMonthRange { get; private set; }
-
-        public ulong MonthFilterGranularity { get; private set; }
-
-        public ulong MaxBytesPerPosition { get; private set; }
-
-        public Optional<ulong> EstimatedAverageBytesPerPosition { get; private set; }
-    }
-
-    public class DatabaseInfo
-    {
-        public DatabaseInfo(string path, bool isOpen)
-        {
-            this.Path = path;
-            this.IsOpen = isOpen;
-            this.Stats = null;
-        }
-
-        public bool IsOpen { get; private set; }
-
-        public string Path { get; private set; }
-
-        public DatabaseStats Stats { get; private set; }
-
-        public void SetStatsFromJson(JObject json)
-        {
-            this.Stats = new DatabaseStats(json);
-        }
-    }
 }
