+++ src/TcecEvaluationBot.ConsoleUI/Services/Models/ChessPosDb/AggregatedEntry.cs
