+++ src/TcecEvaluationBot.ConsoleUI/Services/Models/ChessPosDb/Date.cs
