+++ src/TcecEvaluationBot.ConsoleUI/Services/Models/ChessPosDb/Eco.cs
