+++ src/TcecEvaluationBot.ConsoleUI/Services/Models/ChessPosDb/Entry.cs
