+++ src/TcecEvaluationBot.ConsoleUI/Services/Models/ChessPosDb/GameHeader.cs
