+++ src/TcecEvaluationBot.ConsoleUI/Services/Models/ChessPosDb/GameResult.cs
@@ -1,0 +1,9 @@
+﻿namespace TcecEvaluationBot.ConsoleUI.Services.Models.ChessPosDbQuery
+{
+    public enum GameResult
+    {
+        WhiteWin,
+        BlackWin,
+        Draw,
+    }
+}
