+++ src/TcecEvaluationBot.ConsoleUI/Services/Models/ChessPosDb/MergeMode.cs
@@ -1,0 +1,9 @@
+﻿namespace TcecEvaluationBot.ConsoleUI.Services
+{
+    public enum MergeMode
+    {
+        None,
+        Consecutive,
+        Any,
+    }
+}
