+++ src/TcecEvaluationBot.ConsoleUI/Services/Models/ChessPosDb/Optional.cs
