+++ src/TcecEvaluationBot.ConsoleUI/Services/Models/ChessPosDb/QueryResponse.cs
