+++ src/TcecEvaluationBot.ConsoleUI/Services/Models/ChessPosDb/ResultForRoot.cs
@@ -50,36 +50,4 @@
             return result;
         }
     }
-
-    public class SelectResult
-    {
-        public SelectResult()
-        {
-            this.Children = new Dictionary<string, SegregatedEntries>();
-        }
-
-        public SegregatedEntries Root { get; set; }
-
-        public Dictionary<string, SegregatedEntries> Children { get; set; }
-
-        public static SelectResult FromJson(JObject json)
-        {
-            var result = new SelectResult();
-
-            foreach ((string key, var value) in json)
-            {
-                var entries = SegregatedEntries.FromJson(value.Value<JObject>());
-                if (key == "--")
-                {
-                    result.Root = entries;
-                }
-                else
-                {
-                    result.Children.Add(key, entries);
-                }
-            }
-
-            return result;
-        }
-    }
 }
