+++ src/TcecEvaluationBot.ConsoleUI/Services/Models/ChessPosDb/RootPosition.cs
