+++ src/TcecEvaluationBot.ConsoleUI/Services/Models/ChessPosDb/SegregatedEntries.cs
@@ -58,17 +58,4 @@
             return ((IEnumerable<KeyValuePair<Origin, Entry>>)this.Entries).GetEnumerator();
         }
     }
-
-    internal struct Origin
-    {
-        public Origin(GameLevel level, GameResult result)
-        {
-            this.Level = level;
-            this.Result = result;
-        }
-
-        public GameLevel Level { get; set; }
-
-        public GameResult Result { get; set; }
-    }
 }
