+++ src/TcecEvaluationBot.ConsoleUI/Services/Models/ChessPosDb/Select.cs
@@ -1,0 +1,9 @@
+﻿namespace TcecEvaluationBot.ConsoleUI.Services.Models.ChessPosDbQuery
+{
+    public enum Select
+    {
+        Continuations,
+        Transpositions,
+        All,
+    }
+}
